multimodal-RAG-main/requirements.txt
pinecone
transformers
torch
Pillow
PyMuPDF
google-generativeai
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-magic
passlib[bcrypt]
python-jose[cryptography]
boto3
passlib
